Initial TLS implementation
The Linux side accepts any certificate, but the android side needs
the fingerprint of the linux certificate to trust the computer.
Such fingerprint is shown in a QR code that can be scanned using
the phone camera. More about how the tls trust is made is available
on the wiki:

https://github.com/holylobster/nuntius-android/wiki/TCP-Connection-Use-Case

Note also that for now we rely on an external script to generate
the certificate. This should be replaced in the future by a proper
implementation with GnuTLS.